# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/sync/submissions/{clientRequestId}": {
      "get": {
        "operationId": "SyncController_get",
        "parameters": [
          {
            "name": "X-Store-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado concluído da sincronização.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "submissionId",
                    "clientRequestId",
                    "status",
                    "product",
                    "batch",
                    "initialMovementId",
                    "completedAt"
                  ],
                  "properties": {
                    "submissionId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "clientRequestId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "COMPLETED"
                      ]
                    },
                    "product": {
                      "type": "object",
                      "required": [
                        "id",
                        "created",
                        "version"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "created": {
                          "type": "boolean"
                        },
                        "version": {
                          "type": "integer"
                        }
                      }
                    },
                    "batch": {
                      "type": "object",
                      "required": [
                        "id",
                        "expiryDate",
                        "quantity",
                        "locationId"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "expiryDate": {
                          "type": "string",
                          "format": "date"
                        },
                        "quantity": {
                          "type": "string"
                        },
                        "locationId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    },
                    "initialMovementId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "completedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "retryable",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "LOCATION_NOT_FOUND"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "correlationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "fields": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão para a operação.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "retryable",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "LOCATION_NOT_FOUND"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "correlationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "fields": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Submissão não encontrada no escopo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "retryable",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "LOCATION_NOT_FOUND"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "correlationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "fields": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar resultado de uma sincronização",
        "tags": [
          "sync"
        ]
      }
    },
...
  },
  "info": {
    "title": "ValidaGiro API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}